+++ sampledb/models/locations.py
@@ -39,9 +39,10 @@ class ObjectLocationAssignment(db.Model):
     user_id = db.Column(db.Integer, db.ForeignKey(User.id), nullable=False)
     description = db.Column(db.Text, nullable=False)
     utc_datetime = db.Column(db.DateTime, nullable=False)
+    confirmed = db.Column(db.Boolean, nullable=False, default=False)
     location = db.relationship('Location')
 
-    def __init__(self, object_id: int, location_id: int, user_id: int, description: str, utc_datetime: typing.Optional[datetime.datetime] = None, responsible_user_id: typing.Optional[int] = None):
+    def __init__(self, object_id: int, location_id: int, user_id: int, description: str, utc_datetime: typing.Optional[datetime.datetime] = None, responsible_user_id: typing.Optional[int] = None, confirmed: bool = False):
         self.object_id = object_id
         self.location_id = location_id
         self.user_id = user_id
@@ -50,6 +51,7 @@ class ObjectLocationAssignment(db.Model):
             utc_datetime = datetime.datetime.utcnow()
         self.utc_datetime = utc_datetime
         self.responsible_user_id = responsible_user_id
+        self.confirmed = confirmed
 
     def __repr__(self):
-        return '<{0}(id={1.id}, object_id={1.object_id}, location_id={1.location_id}, user_id={1.user_id}, responsible_user_id={1.responsible_user_id}, utc_datetime={1.utc_datetime}, description="{1.content}")>'.format(type(self).__name__, self)
+        return '<{0}(id={1.id}, object_id={1.object_id}, location_id={1.location_id}, user_id={1.user_id}, responsible_user_id={1.responsible_user_id}, utc_datetime={1.utc_datetime}, description="{1.content}", confirmed={1.confirmed)>'.format(type(self).__name__, self)